+++ cli/src/operation_templater.rs
@@ -40,6 +40,7 @@ use crate::template_builder::merge_fn_map;
 use crate::template_parser;
 use crate::template_parser::FunctionCallNode;
 use crate::template_parser::TemplateDiagnostics;
+use crate::template_parser::TemplateParseError;
 use crate::template_parser::TemplateParseResult;
 use crate::templater::BoxedAnyProperty;
 use crate::templater::BoxedSerializeProperty;
@@ -522,13 +523,33 @@ where
         },
     );
     map.insert(
-        "tags",
+        "attributes",
         |_language, _diagnostics, _build_ctx, self_property, function| {
             function.expect_no_arguments()?;
             let out_property = self_property.map(|op| {
                 // TODO: introduce map type
                 op.metadata()
-                    .tags
+                    .attributes
+                    .iter()
+                    .map(|(key, value)| format!("{key}: {value}"))
+                    .join("\n")
+            });
+            Ok(out_property.into_dyn_wrapped())
+        },
+    );
+    // TODO: Remove in jj 0.47+
+    map.insert(
+        "tags",
+        |_language, diagnostics, _build_ctx, self_property, function| {
+            diagnostics.add_warning(TemplateParseError::expression(
+                "operation.tags() is deprecated; use .attributes() instead",
+                function.name_span,
+            ));
+            function.expect_no_arguments()?;
+            let out_property = self_property.map(|op| {
+                // TODO: introduce map type
+                op.metadata()
+                    .attributes
                     .iter()
                     .map(|(key, value)| format!("{key}: {value}"))
                     .join("\n")