+++ lib/src/protos/simple_op_store.rs
@@ -189,7 +189,7 @@ pub struct OperationMetadata {
     #[prost(string, optional, tag = "8")]
     pub workspace_name: ::core::option::Option<::prost::alloc::string::String>,
     #[prost(map = "string, string", tag = "6")]
-    pub tags: ::std::collections::HashMap<
+    pub attributes: ::std::collections::HashMap<
         ::prost::alloc::string::String,
         ::prost::alloc::string::String,
     >,